mv to series (WIP)
