+++ packages/perspective-viewer-d3fc/src/js/series/xy-scatter/xyScatterSeries.js
@@ -1,0 +1,173 @@
+// ┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┓
+// ┃ ██████ ██████ ██████       █      █      █      █      █ █▄  ▀███ █       ┃
+// ┃ ▄▄▄▄▄█ █▄▄▄▄▄ ▄▄▄▄▄█  ▀▀▀▀▀█▀▀▀▀▀ █ ▀▀▀▀▀█ ████████▌▐███ ███▄  ▀█ █ ▀▀▀▀▀ ┃
+// ┃ █▀▀▀▀▀ █▀▀▀▀▀ █▀██▀▀ ▄▄▄▄▄ █ ▄▄▄▄▄█ ▄▄▄▄▄█ ████████▌▐███ █████▄   █ ▄▄▄▄▄ ┃
+// ┃ █      ██████ █  ▀█▄       █ ██████      █      ███▌▐███ ███████▄ █       ┃
+// ┣━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┫
+// ┃ Copyright (c) 2017, the Perspective Authors.                              ┃
+// ┃ ╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌ ┃
+// ┃ This file is part of the Perspective library, distributed under the terms ┃
+// ┃ of the [Apache License 2.0](https://www.apache.org/licenses/LICENSE-2.0). ┃
+// ┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┛
+
+import * as fc from "d3fc";
+import { axisFactory } from "../../axis/axisFactory";
+import { chartCanvasFactory } from "../../axis/chartFactory";
+import {
+    pointSeriesCanvas,
+    symbolTypeFromColumn,
+} from "../../series/pointSeriesCanvas";
+import { seriesLinearRange } from "../../series/seriesRange";
+import withGridLines from "../../gridlines/gridlines";
+import { hardLimitZeroPadding } from "../../d3fc/padding/hardLimitZero";
+import zoomableChart from "../../zoom/zoomableChart";
+import nearbyTip from "../../tooltip/nearbyTip";
+import { symbolsObj } from "../../series/seriesSymbols";
+
+/**
+ * Define a clamped scaling factor based on the container size for bubble plots.
+ *
+ * @param {Array} p1 a point as a tuple of `Number`
+ * @param {Array} p2 a second point as a tuple of `Number`
+ * @returns a function `container -> integer` which calculates a scaling factor
+ * from the linear function (clamped) defgined by the input points
+ */
+function interpolate_scale([x1, y1], [x2, y2]) {
+    const m = (y2 - y1) / (x2 - x1);
+    const b = y2 - m * x2;
+    return function (container) {
+        const node = container.node();
+        const shortest_axis = Math.min(node.clientWidth, node.clientHeight);
+        return Math.min(y2, Math.max(y1, m * shortest_axis + b));
+    };
+}
+
+/**
+ * @param {d3.Selection} container - d3.Selection of the outer div
+ * @param {any} settings - settings as defined in the Update method in plugin.js
+ */
+export default function xyScatterSeries() {
+    let settings = null;
+    let data = null;
+    let color = null;
+    let symbols = null;
+
+    const _xyScatterSeries = (container) => {
+        const colorBy = settings.realValues[2];
+        let hasColorBy = !!colorBy;
+        const symbolCol = settings.realValues[4];
+
+        const size = settings.realValues[3]
+            ? seriesLinearRange(settings, data, "size").range([10, 10000])
+            : null;
+
+        const label = settings.realValues[5];
+
+        const scale_factor = interpolate_scale(
+            [600, 0.1],
+            [1600, 1]
+        )(container);
+        const series = fc
+            .seriesCanvasMulti()
+            .mapping((data, index) => data[index])
+            .series(
+                data.map((series) =>
+                    pointSeriesCanvas(
+                        settings,
+                        symbolCol,
+                        size,
+                        color,
+                        label,
+                        symbols,
+                        scale_factor
+                    )
+                )
+            );
+
+        const axisDefault = () =>
+            axisFactory(settings)
+                .settingName("mainValues")
+                .paddingStrategy(hardLimitZeroPadding())
+                .pad([0.1, 0.1]);
+
+        const xAxis = axisDefault()
+            .settingValue(settings.mainValues[0].name)
+            .memoValue(settings.axisMemo[0])
+            .valueName("x")(data);
+
+        const yAxis = axisDefault()
+            .orient("vertical")
+            .settingValue(settings.mainValues[1].name)
+            .memoValue(settings.axisMemo[1])
+            .valueName("y")(data);
+
+        const chart = chartCanvasFactory(xAxis, yAxis)
+            .xLabel("")
+            .yLabel("")
+            .plotArea(withGridLines(series, settings).canvas(true));
+
+        chart.xNice && chart.xNice();
+        chart.yNice && chart.yNice();
+
+        // TODO: This make scrolling the grid tricky. Maybe require modifier key to zoom?
+        // Also, reset zoom button is now at the top of the page and overlaps any other items.
+        const zoomChart = zoomableChart()
+            .chart(chart)
+            .settings(settings)
+            .xScale(xAxis.scale)
+            .yScale(yAxis.scale)
+            .canvas(true);
+
+        const toolTip = nearbyTip()
+            .scaleFactor(scale_factor)
+            .settings(settings)
+            .canvas(true)
+            .xScale(xAxis.scale)
+            .xValueName("x")
+            .yValueName("y")
+            .yScale(yAxis.scale)
+            .color(!hasColorBy && color)
+            .size(size)
+            .data(data);
+
+        // render
+        container.datum(data.map((d) => [d])).call(zoomChart);
+        container.call(toolTip);
+    };
+
+    _xyScatterSeries.settings = (...args) => {
+        if (!args.length) {
+            return settings;
+        }
+        settings = args[0];
+        return _xyScatterSeries;
+    };
+    _xyScatterSeries.data = (...args) => {
+        if (!args.length) {
+            return data;
+        }
+        data = args[0];
+        return _xyScatterSeries;
+    };
+    _xyScatterSeries.color = (...args) => {
+        if (!args.length) {
+            return color;
+        }
+        color = args[0];
+        return _xyScatterSeries;
+    };
+    _xyScatterSeries.symbols = (...args) => {
+        if (!args.length) {
+            return symbols;
+        }
+        symbols = args[0];
+        return _xyScatterSeries;
+    };
+
+    // mk_get_set("settings", settings);
+    // mk_get_set("data", data);
+    // mk_get_set("color", color);
+    // mk_get_set("symbols", symbols);
+
+    return _xyScatterSeries;
+}
